Amodal-Completion-in-the-Wild/requirements.txt
opencv-python>=2.4.9
pyyaml
imutils
cvbase
tensorboardX
networkx
cvbase
scikit-image
pillow
tqdm
pycocotools
ipdb
diffusers==0.15.0
huggingface-hub==0.23.2
accelerate>=0.26.0
xformers==0.0.23.post1